Added common implementation of 'iterator' and macros for member type declarations
As many containers share implementation details of iterator and declarations of member types (size_type, value_type, refence ...)
it seemed logical to place those in a sepearate file. The 'iterator_impl' template mainly just defines the common operators of iterators (like ++, *, ->).
It forwards implementation specific functions to the per container specialized iterators, that are passed as a template argument to 'iterator_impl'.